Add ruff config, CI workflow, and fix stale test-count doc
Adds [tool.ruff] to pyproject.toml (line-length 120, E/F/I/UP/B/RUF rules)
plus a .github/workflows/ci.yml running ruff + pytest on push/PR. No lint
enforcement existed despite an orphaned .ruff_cache/. Autofixes typing
cleanups (quoted annotations, Iterable/Mapping -> collections.abc) and two
manual fixes (unused unpacked variable, redundant int() cast). README test
count corrected from a stale 172 to the actual 391.

+++ src/postureguard/ui/screens/settings.py
@@ -642,7 +642,7 @@ class SettingsScreen(QWidget):
         """Reflect a sensitivity change made outside this screen — the automatic
         backoff after repeated snoozes — without re-emitting `changed` for it."""
         self._loading = True
-        self.sensitivity.slider.setValue(int(round(value * 100)))
+        self.sensitivity.slider.setValue(round(value * 100))
         self._loading = False
 
     def refresh(self) -> None:
